packages/database/pyproject.toml
[project]
name = "database"
version = "0.1.0"
description = "Shared database models and schemas"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "pgvector",
    "alembic",
    "asyncpg",
    "psycopg2-binary",
    "redis",
    "python-dotenv",
    "utils",
]

[tool.uv.sources]
utils = { workspace = true }
